+++ inc/Options/Wizard.php
@@ -87,6 +87,10 @@ class Wizard {
 				return new Fields\UrlField( $id, $title, $args_without_type );
 			case 'encrypted_password':
 				return new Fields\EncrypytedPasswordField( $id, $title, $args_without_type );
+			case 'hidden':
+				return new Fields\HiddenField( $id, $title, $args_without_type );
+			case 'image':
+				return new Fields\ImageField( $id, $title, $args_without_type );
 			default:
 				throw new \Exception( "Unrecognized type '{$args['type']}'" );
 		}